chore: remove token list item v2 flag
