+++ app/components/UI/TokenDetails/Views/TokenDetails.tsx
@@ -1,7 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useRef } from 'react';
 import { View, StyleSheet, ActivityIndicator } from 'react-native';
 import { useSelector } from 'react-redux';
-import { selectTokenListLayoutV2Enabled } from '../../../../selectors/featureFlagController/tokenListLayout';
 import { useAnalytics } from '../../../hooks/useAnalytics/useAnalytics';
 import { MetaMetricsEvents } from '../../../../core/Analytics';
 import { SupportedCaipChainId } from '@metamask/multichain-network-controller';
@@ -317,7 +316,6 @@ const TokenDetails: React.FC<{
 const useTokenDetailsOpenedTracking = (params: TokenDetailsRouteParams) => {
   const { trackEvent, createEventBuilder } = useAnalytics();
   const { variantName, isTestActive } = useTokenDetailsABTest();
-  const isTokenListV2 = useSelector(selectTokenListLayoutV2Enabled);
   const lastTrackedTokenKeyRef = useRef<string | null>(null);
 
   return useCallback(
@@ -356,16 +354,9 @@ const useTokenDetailsOpenedTracking = (params: TokenDetailsRouteParams) => {
         market_insights_displayed: isMarketInsightsDisplayed,
         severity,
         // A/B test attribution — each experiment is independent
-        ...((isTestActive || isFromTokenList) && {
+        ...(isTestActive && {
           ab_tests: {
-            ...(isTestActive && {
-              assetsASSETS2493AbtestTokenDetailsLayout: variantName,
-            }),
-            ...(isFromTokenList && {
-              assetsASSETS2621AbtestTokenListLayout: isTokenListV2
-                ? 'v2'
-                : 'v1',
-            }),
+            assetsASSETS2493AbtestTokenDetailsLayout: variantName,
           },
         }),
       };
@@ -378,7 +369,6 @@ const useTokenDetailsOpenedTracking = (params: TokenDetailsRouteParams) => {
     [
       createEventBuilder,
       isTestActive,
-      isTokenListV2,
       params.address,
       params.balance,
       params.chainId,
