+++ app/components/UI/Tokens/TokenList/TokenList.tsx
@@ -17,7 +17,6 @@ import {
 import { TokenI } from '../types';
 import { strings } from '../../../../../locales/i18n';
 import { TokenListItem } from './TokenListItem/TokenListItem';
-import { TokenListItemV2 } from './TokenListItemV2/TokenListItemV2';
 import { WalletViewSelectorsIDs } from '../../../Views/Wallet/WalletView.testIds';
 import { useNavigation } from '@react-navigation/native';
 import Routes from '../../../../constants/navigation/Routes';
@@ -31,7 +30,6 @@ import { useTailwind } from '@metamask/design-system-twrnc-preset';
 import { MetaMetricsEvents } from '../../../../core/Analytics';
 import { useAnalytics } from '../../../hooks/useAnalytics/useAnalytics';
 import { SCROLL_TO_TOKEN_EVENT } from '../constants';
-import { selectTokenListLayoutV2Enabled } from '../../../../selectors/featureFlagController/tokenListLayout';
 import { useMusdCtaVisibility } from '../../Earn/hooks/useMusdCtaVisibility';
 
 export interface FlashListAssetKey {
@@ -73,10 +71,6 @@ const TokenListComponent = ({
 
   // Declaring this here and passing it down to avoid O(n) API calls to on-ramp
   const { shouldShowTokenListItemCta } = useMusdCtaVisibility();
-
-  // A/B test: Token list item layout (V1 vs V2)
-  const isTokenListV2 = useSelector(selectTokenListLayoutV2Enabled);
-  const ListItemComponent = isTokenListV2 ? TokenListItemV2 : TokenListItem;
 
   const listRef = useRef<FlashListRef<FlashListAssetKey>>(null);
 
@@ -166,7 +160,7 @@ const TokenListComponent = ({
 
   const renderTokenListItem = useCallback(
     ({ item }: { item: FlashListAssetKey }) => (
-      <ListItemComponent
+      <TokenListItem
         assetKey={item}
         showRemoveMenu={showRemoveMenu}
         setShowScamWarningModal={setShowScamWarningModal}
@@ -178,7 +172,6 @@ const TokenListComponent = ({
       />
     ),
     [
-      ListItemComponent,
       showRemoveMenu,
       setShowScamWarningModal,
       privacyMode,
@@ -197,7 +190,7 @@ const TokenListComponent = ({
         testID={WalletViewSelectorsIDs.TOKENS_CONTAINER_LIST}
       >
         {displayTokenKeys.map((item, index) => (
-          <ListItemComponent
+          <TokenListItem
             key={`${getTokenKey(item)}-${index}`}
             assetKey={item}
             showRemoveMenu={showRemoveMenu}
